Create WebApp Blank Code - NO CRED INCL
The code shared here features the webapp code to take user data and run inference on your Roboflow server. The output is shown as an aggregated bar graph of the object detection from all user images. The code also tests the image exif for geotag data and displays available information on a map.